+++ csvkit/utilities/csvjson.py
@@ -37,7 +37,7 @@ class CSVJSON(CSVKitUtility):
         self.argparser.add_argument('-y', '--snifflimit', dest='sniff_limit', type=int,
                                     help='Limit CSV dialect sniffing to the specified number of bytes. Specify "0" to disable sniffing entirely.')
         self.argparser.add_argument('-I', '--no-inference', dest='no_inference', action='store_true',
-                                    help='Disable type inference when parsing CSV input.')
+                                    help='Disable type inference (and --locale, --date-format, --datetime-format) when parsing CSV input.')
 
     def main(self):
         # We need to do this dance here, because we aren't writing through agate.
@@ -160,7 +160,7 @@ class CSVJSON(CSVKitUtility):
                 ])
 
             dump_json(output)
-        elif self.args.streamOutput and self.args.no_inference:
+        elif self.args.streamOutput and self.args.no_inference and not self.args.skip_lines:
             rows = agate.csv.reader(self.input_file, **self.reader_kwargs)
             column_names = next(rows)